bunch of cleanup and doc comments
