+++ core/src/data/midi_message.rs
@@ -60,7 +60,7 @@ impl Display for MidiMessage {
             MidiMessage::NoteOff { key, velocity } => {
                 f.write_str(&format!("[note_off]key:{}, velocity:{}", key, velocity))
             }
-            MidiMessage::AllNoteOff => f.write_str(&format!("[global_note_off]")),
+            MidiMessage::AllNoteOff => f.write_str("[global_note_off]"),
         }
     }
 }
