+++ gui/src/ui.rs
@@ -1,5 +1,7 @@
 /// common types
 pub mod common;
+
+/// ui elements
 mod element;
 
 pub use element::*;
