+++ gui/src/ui/element/note_editor.rs
@@ -2,215 +2,20 @@ use crate::EDGE_COLOR;
 use egui::{
     emath::Real,
     epaint::{self, CircleShape},
-    lerp, pos2, vec2, Color32, FontId, Id, LayerId, Order, PointerButton, Pos2, Rect, Response,
-    Rounding, Sense, Shape, Stroke, Ui, Vec2,
+    lerp, pos2, Color32, FontId, Id, LayerId, Order, PointerButton, Pos2, Rect, Response, Rounding,
+    Sense, Shape, Stroke, Ui,
 };
+
+use self::transform::Transform;
 use hexencer_core::{data::Clip, Tick};
 use std::ops::RangeInclusive;
 
 use super::BEAT_WIDTH;
 
-const EDITOR_NOTE_HEIGHT: f32 = 8.0;
+/// trainsform used for panning and zooming
+mod transform;
 
-#[derive(Clone, Copy)]
-struct EditorBounds {
-    pub min: [f64; 2],
-    pub max: [f64; 2],
-}
-impl EditorBounds {
-    fn to_rect(&self) -> Rect {
-        Rect::from_min_max(
-            pos2(self.min[0] as f32, self.min[1] as f32),
-            pos2(self.max[0] as f32, self.max[1] as f32),
-        )
-    }
-
-    pub const NONE: Self = Self {
-        min: [f64::INFINITY; 2],
-        max: [f64::INFINITY; 2],
-    };
-
-    fn box_zoom(&mut self, zoom_factor: Vec2, center: EditorPoint) {
-        self.min[0] = center.x + (self.min[0] - center.x) / (zoom_factor.x as f64);
-        self.max[0] = center.x + (self.max[0] - center.x) / (zoom_factor.x as f64);
-        self.min[1] = center.y + (self.min[1] - center.y) / (zoom_factor.y as f64);
-        self.max[1] = center.y + (self.max[1] - center.y) / (zoom_factor.y as f64);
-    }
-
-    pub fn is_finite(&self) -> bool {
-        self.min[0].is_finite()
-            && self.min[1].is_finite()
-            && self.max[0].is_finite()
-            && self.max[1].is_finite()
-    }
-
-    fn is_valid(&self) -> bool {
-        self.is_finite() && self.width() > 0.0 && self.height() > 0.0
-    }
-
-    fn width(&self) -> f64 {
-        self.max[0] - self.min[0]
-    }
-
-    fn height(&self) -> f64 {
-        self.max[1] - self.min[1]
-    }
-}
-
-#[derive(Clone, Copy)]
-pub struct Transform {
-    bounds: EditorBounds,
-    scale: f32,
-    frame: Rect,
-    translation: Vec2,
-}
-
-#[derive(Clone, Debug)]
-struct EditorPoint {
-    pub x: f64,
-    pub y: f64,
-}
-
-impl EditorPoint {
-    fn new(x: impl Into<f64>, y: impl Into<f64>) -> Self {
-        Self {
-            x: x.into(),
-            y: y.into(),
-        }
-    }
-}
-
-impl Transform {
-    fn zoom(
-        &mut self,
-        zoom_delta: Vec2,
-        screen_hover_pos: Pos2,
-        ui: &mut Ui,
-        rect: Rect,
-        step_size: f32,
-    ) {
-        let min_height = rect.min.y;
-        let max_height = rect.max.y;
-
-        let zoom_rate = 0.002;
-        let zoom_delta_scaled_by_rate = zoom_delta * zoom_rate;
-        let clamped_scale = (self.scale + zoom_delta_scaled_by_rate.x).clamp(0.8, 4.0);
-
-        let pointer_in_editor_space = (screen_hover_pos - self.translation) / self.scale;
-        self.scale = clamped_scale;
-
-        let mut new_translation = screen_hover_pos - (pointer_in_editor_space * self.scale);
-        new_translation.y += zoom_delta.y * self.scale;
-
-        let total_height = step_size * 128.0;
-        if new_translation.y >= min_height {
-            new_translation.y = min_height;
-        } else if new_translation.y + (total_height * self.scale) < max_height {
-            new_translation.y = max_height - (total_height * self.scale);
-        }
-        self.translation = new_translation;
-
-        if cfg!(feature = "debug") {
-            self.debug_lines(
-                pointer_in_editor_space,
-                new_translation,
-                ui,
-                screen_hover_pos,
-                total_height,
-            );
-        }
-    }
-
-    fn debug_lines(
-        &mut self,
-        pointer_in_editor_space: Pos2,
-        new_translation: Vec2,
-        ui: &mut Ui,
-        screen_hover_pos: Pos2,
-        total_height: f32,
-    ) {
-        let pointer_origin_scaled = pos2(125.0, 0.0);
-        let pointer_translation_scaled =
-            pos2(125.0, new_translation.y + (total_height * self.scale));
-        debug_line(
-            ui,
-            pointer_origin_scaled,
-            pointer_translation_scaled,
-            Color32::LIGHT_YELLOW,
-        );
-        let pointer_in_editor_space_clamped = pos2(1.0, pointer_in_editor_space.y);
-        let scaled = pointer_in_editor_space * self.scale;
-        let scaled_clamped = pos2(1.0, scaled.y);
-        let new_center_clamped = pos2(10.0, new_translation.y);
-        debug_dot(ui, pointer_in_editor_space_clamped, Color32::RED);
-        debug_dot(ui, scaled_clamped, Color32::LIGHT_BLUE);
-        debug_dot(ui, new_center_clamped, Color32::GREEN);
-        let editor_origin = pos2(150.0, 0.0);
-        let editor_translation = pos2(150.0, self.translation.y);
-        debug_line(ui, editor_origin, editor_translation, Color32::RED);
-        let pointer_origin_before_scaling = pos2(200.0, 0.0);
-        let pointer_translation_before_scaling = pos2(200.0, pointer_in_editor_space.y);
-        debug_line(
-            ui,
-            pointer_origin_before_scaling,
-            pointer_translation_before_scaling,
-            Color32::GREEN,
-        );
-        let pointer_origin_scaled = pos2(125.0, 0.0);
-        let pointer_translation_scaled = pos2(125.0, pointer_in_editor_space.y * self.scale);
-        debug_line(
-            ui,
-            pointer_origin_scaled,
-            pointer_translation_scaled,
-            Color32::LIGHT_YELLOW,
-        );
-        let transform_offset_origin = pos2(175.0, self.translation.y);
-        let transform_offset = pos2(
-            175.0,
-            screen_hover_pos.y + (pointer_in_editor_space.y * self.scale),
-        );
-        debug_line(
-            ui,
-            transform_offset_origin,
-            transform_offset,
-            Color32::from_rgb(255, 0, 255),
-        );
-    }
-
-    /// Convert a position in the frame to a value in the bounds.
-    fn value_from_position(&self, pos: Pos2) -> EditorPoint {
-        tracing::info!("frame {:?}", self.frame);
-        let x = remap(
-            pos.x as f64,
-            (self.frame.left() as f64)..=(self.frame.right() as f64),
-            self.bounds.min[0]..=self.bounds.max[0],
-        );
-        let y = remap(
-            pos.y as f64,
-            (self.frame.bottom() as f64)..=(self.frame.top() as f64),
-            self.bounds.min[1]..=self.bounds.max[1],
-        );
-        EditorPoint::new(x, y)
-    }
-
-    fn new(frame: Rect, bounds: EditorBounds, scale: f32) -> Self {
-        Self {
-            frame,
-            bounds,
-            scale,
-            translation: frame.min.to_vec2(),
-        }
-    }
-
-    fn apply(&self, point: Pos2) -> Pos2 {
-        point + self.translation
-    }
-
-    fn inverse_apply(&self, point: Pos2) -> Pos2 {
-        point - self.translation
-    }
-}
-
+/// paint a debug line on the top layer
 fn debug_line(ui: &mut Ui, origin: Pos2, target: Pos2, color: Color32) {
     let top_layer = LayerId::new(Order::Foreground, Id::new("debug_line"));
     ui.with_layer_id(top_layer, |ui| {
@@ -219,6 +24,7 @@ fn debug_line(ui: &mut Ui, origin: Pos2, target: Pos2, color: Color32) {
     });
 }
 
+/// paint a debug dot on top layer
 fn debug_dot(ui: &Ui, point: Pos2, color: Color32) {
     ui.painter().add(CircleShape {
         center: point,
@@ -240,32 +46,42 @@ where
     lerp(to, t)
 }
 
+/// note editor state
 #[derive(Clone)]
 pub struct State {
+    /// the last click position when zooming in the note editor
     pub last_click_pos_for_zoom: Option<Pos2>,
+    /// transform state for the note editor, used for zooming and panning
     pub transform: Transform,
+    /// step size for the note editor
     pub step_size: f32,
 }
 
 impl State {
+    /// load the note editor state from memory
     pub fn load(ui: &Ui, id: Id) -> Option<Self> {
         ui.memory(|memory| memory.data.get_temp(id))
     }
 
+    /// store the note editor state to memory
     pub fn store(self, ui: &Ui, id: Id) {
         ui.memory_mut(|memory| memory.data.insert_temp(id, self));
     }
 }
 
+/// A note editor widget that can be used to display and edit notes of the currently selected clip
 pub struct NoteEditorWidget<'c> {
+    /// reference to the clip this note editor is editing
     clip: &'c Clip,
 }
 
 impl<'c> NoteEditorWidget<'c> {
+    /// Create a new note editor widget
     pub fn new(clip: &'c Clip) -> Self {
         Self { clip }
     }
 
+    /// setup and render the widget, and return a response
     pub fn show(self, ui: &mut Ui) -> Response {
         let fill_color = Color32::from_rgb(50, 50, 50);
         let editor_rect = ui.available_rect_before_wrap();
@@ -273,10 +89,9 @@ impl<'c> NoteEditorWidget<'c> {
         ui.set_clip_rect(editor_rect);
         let id = Id::new("note_editor");
 
-        let min_auto_bounds = EditorBounds::NONE;
         let mut state = State::load(ui, id).unwrap_or_else(|| State {
             last_click_pos_for_zoom: None,
-            transform: Transform::new(editor_rect, min_auto_bounds, 1.0),
+            transform: Transform::new(editor_rect, 1.0),
             step_size: 10.0,
         });
         let shape = epaint::RectShape::new(
@@ -326,6 +141,7 @@ impl<'c> NoteEditorWidget<'c> {
         response
     }
 
+    /// draw the note lanes in the editor
     fn draw_note_lanes(
         &self,
         ui: &mut Ui,
@@ -365,7 +181,7 @@ impl<'c> NoteEditorWidget<'c> {
                     ui,
                     scaled_step_size,
                     transform,
-                    &tick,
+                    tick,
                     editor_rect,
                 );
             }
@@ -374,6 +190,7 @@ impl<'c> NoteEditorWidget<'c> {
         ui.painter().extend(lines);
     }
 
+    /// draws the beat columns in the note editor
     fn draw_beat_columns(&self, ui: &mut Ui, rect: Rect) {
         let mut shapes = Vec::new();
         for i in 0..100 {
@@ -388,15 +205,16 @@ impl<'c> NoteEditorWidget<'c> {
     }
 }
 
+/// paints the note editor line numbers
 fn line_number(ui: &mut Ui, pos: Pos2, num: i32) {
     let font_id = FontId::monospace(12.0);
     let text_color = Color32::RED;
     let galley = ui.fonts(|f| {
         f.layout(
-            String::from({
+            {
                 let output_text = format!("note {}", num);
                 output_text
-            }),
+            },
             font_id,
             text_color,
             10000.0,
@@ -415,7 +233,7 @@ fn line_number(ui: &mut Ui, pos: Pos2, num: i32) {
     });
 }
 
-// this is box zooming, draw a rectangle around the area you want to zoom into to focus in on it
+/// this is box zooming, draw a rectangle around the area you want to zoom into to focus in on it
 fn box_zooming(
     response: &Response,
     state: &mut State,
@@ -454,11 +272,7 @@ fn box_zooming(
     }
 }
 
-#[derive(Clone, Debug)]
-struct GridUnit {
-    pub offset: f64,
-}
-
+/// create a clip note
 fn clip_note(
     note_lane: u8,
     ui: &mut Ui,
